complete shared library and update dialog box usage
