+++ client/package.json
@@ -95,7 +95,8 @@
     "electron-updater": "^6.1.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.16.0"
+    "react-router-dom": "^6.16.0",
+    "syncstream-sharedlib": "file:../sharedLib"
   },
   "devDependencies": {
     "@electron/rebuild": "^3.3.0",
