+++ client/tsconfig.json
@@ -13,14 +13,9 @@
     "resolveJsonModule": true,
     "allowJs": true,
     "outDir": ".erb/dll",
-    "paths": {
-      "@types/*": ["../shared/types/*"],
-      "shared": ["../shared"]
-    },
     "typeRoots": [
-      "../shared/types",
       "./node_modules/@types",
-    ]
+    ],
   },
   "exclude": ["test", "release/build", "release/app/dist", ".erb/dll"]
 }
