+++ server/tsconfig.json
@@ -27,12 +27,10 @@
     // "moduleResolution": "Node",                     /* Specify how TypeScript looks up a file from a given module specifier. */
     "baseUrl": ".",                                  /* Specify the base directory to resolve non-relative module names. */
     "paths": {                                           /* Specify a set of entries that re-map imports to additional lookup locations. */
-      "@types/*": ["./types/*", "../shared/types/*"],
-      "shared": ["../shared"]
+      "@types/*": ["./types/*"],
     },                                                   
     // "rootDirs": [],                                   /* Allow multiple folders to be treated as one when resolving modules. */
     "typeRoots": [
-      "../shared/types",
       "./types",
       "./node_modules/@types"
     ],                                  /* Specify multiple folders that act like './node_modules/@types'. */
@@ -105,8 +103,13 @@
     // "allowUnreachableCode": true,                     /* Disable error reporting for unreachable code. */
     /* Completeness */
     // "skipDefaultLibCheck": true,                      /* Skip type checking .d.ts files that are included with TypeScript. */
-    "skipLibCheck": true /* Skip type checking all .d.ts files. */
+    "skipLibCheck": true, /* Skip type checking all .d.ts files. */
   },
-  "include": ["src", "types", "../shared/**/*"],
+  "references": [
+    { 
+      "path": "../sharedLib",
+    }
+  ],
+  "include": ["src", "types"],
   "exclude": ["src/**/*.test.ts"]
 }
