--- shared/index.ts
@@ -1,2 +1,0 @@
-export * as Types from "./types"
-export * as Validation from "./validation"
