--- shared/validation/index.ts
@@ -1,1 +1,0 @@
-export * from "./validate_types"
