+++ sharedLib/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "syncstream-sharedlib",
+  "version": "0.0.0",
+  "description": "Shared library for common code between client and server.",
+  "main": "index.js",
+  "scripts": {
+    "build": "tsc"
+  },
+  "author": "",
+  "license": "MIT",
+  "devDependencies": {
+    "@types/node": "^22.13.4",
+    "typescript": "^5.7.3"
+  }
+}
