+++ sharedLib/tsconfig.json
@@ -1,0 +1,12 @@
+// lib/tsconfig.json
+{
+  "compilerOptions": {
+    /* Truncated compiler options to list only relevant options */
+    "declaration": true, 
+    "declarationMap": true,
+    "rootDir": ".",   
+    "composite": true,
+    "outDir": "./dist",
+  },
+  "references": []  // * Any project that is referenced must itself have a `references` array (which may be empty).
+}
