+++ sharedLib/types/index.ts
@@ -1,0 +1,3 @@
+export * from './misc'
+export * from './room'
+export * from './user'
