+++ sharedLib/types/misc.ts
@@ -1,0 +1,4 @@
+// Used for basic communication
+export interface StringMessage {
+  msg: string;
+}
