+++ sharedLib/types/room.ts
@@ -1,24 +1,3 @@
-// MISC
-export interface StringMessage {
-  msg: string;
-}
-
-// USER
-export interface UserData {
-  username: string;
-  email?: string;
-  password?: string;
-  admin?: boolean;
-  displayName?: string;
-}
-
-export interface UserUpdateData {
-  email?: string;
-  password?: string;
-  displayName?: string;
-}
-
-// ROOM
 export interface RoomPermissions {
   msg: string;
   // TODO: Complete room permissions setup
