+++ sharedLib/validation/misc.ts
@@ -1,0 +1,5 @@
+// Misc type validation
+
+export function isValidStringMessage(sm: any): boolean {
+  return !!sm && typeof sm.msg === "string"
+}
